+++ packages/rekapager-doctrine-orm-adapter/src/QueryBuilderAdapter.php
@@ -18,6 +18,7 @@ use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\DBAL\Types\Types;
+use Doctrine\ORM\Query\Expr\Andx;
 use Doctrine\ORM\Query\Expr\From;
 use Doctrine\ORM\Query\Expr\OrderBy;
 use Doctrine\ORM\QueryBuilder;
@@ -27,6 +28,7 @@ use Rekalogika\Contracts\Rekapager\Exception\UnexpectedValueException;
 use Rekalogika\Rekapager\Adapter\Common\Field;
 use Rekalogika\Rekapager\Adapter\Common\IndexResolver;
 use Rekalogika\Rekapager\Adapter\Common\KeysetExpressionCalculator;
+use Rekalogika\Rekapager\Adapter\Common\SeekMethod;
 use Rekalogika\Rekapager\Doctrine\ORM\Exception\UnsupportedQueryBuilderException;
 use Rekalogika\Rekapager\Doctrine\ORM\Internal\KeysetQueryBuilderVisitor;
 use Rekalogika\Rekapager\Doctrine\ORM\Internal\QueryBuilderKeysetItem;
@@ -61,6 +63,7 @@ final class QueryBuilderAdapter implements KeysetPaginationAdapterInterface, Off
         private readonly array $typeMapping = [],
         private readonly bool|null $useOutputWalkers = null,
         private readonly string|null $indexBy = null,
+        private readonly SeekMethod $seekMethod = SeekMethod::Approximated,
     ) {
         if ($queryBuilder->getFirstResult() !== 0 || $queryBuilder->getMaxResults() !== null) {
             throw new UnsupportedQueryBuilderException();
@@ -139,20 +142,14 @@ final class QueryBuilderAdapter implements KeysetPaginationAdapterInterface, Off
 
         // returns early if there are no boundary values
 
-        $fields = $this->createCalculatorFields($boundaryValues, $orderings);
+        [$where, $parameters] = $this->generateWhereExpression(
+            boundaryValues: $boundaryValues,
+            orderings: $orderings
+        );
 
-        if ($fields === []) {
-            return $queryBuilder;
-        }
+        $queryBuilder->andWhere($where);
 
-        // adds where expression to the querybuilder
-
-        $keysetExpression = KeysetExpressionCalculator::calculate($fields);
-
-        $visitor = new KeysetQueryBuilderVisitor();
-        $queryBuilder->andWhere($visitor->dispatch($keysetExpression));
-
-        foreach ($visitor->getParameters() as $template => $parameter) {
+        foreach ($parameters as $template => $parameter) {
             $queryBuilder->setParameter(
                 $template,
                 $parameter->getValue(),
@@ -161,6 +158,46 @@ final class QueryBuilderAdapter implements KeysetPaginationAdapterInterface, Off
         }
 
         return $queryBuilder;
+    }
+
+    /**
+     * @param array<string,QueryParameter> $boundaryValues Key is the property name, value is the bound value. Null if unbounded.
+     * @param non-empty-array<string,'ASC'|'DESC'> $orderings
+     * @return array{?Andx,array<string,QueryParameter>}
+     */
+    private function generateWhereExpression(
+        array $boundaryValues,
+        array $orderings,
+    ): array {
+        $fields = $this->createCalculatorFields($boundaryValues, $orderings);
+
+        if ($fields === []) {
+            return [null, []];
+        }
+
+        return match ($this->seekMethod) {
+            SeekMethod::Approximated => $this->generateApproximatedWhereExpression($fields),
+                // SeekMethod::RowValues => $this->generateRowValuesWhereExpression($fields),
+                // SeekMethod::Auto => $this->generateAutoWhereExpression($fields),
+            default => throw new LogicException('Unsupported seek method'),
+        };
+    }
+
+    /**
+     * @param non-empty-list<Field> $fields
+     * @return array{?Andx,array<string,QueryParameter>}
+     */
+    private function generateApproximatedWhereExpression(array $fields): array
+    {
+        $expression = KeysetExpressionCalculator::calculate($fields);
+
+        $visitor = new KeysetQueryBuilderVisitor();
+        $where = $visitor->dispatch($expression);
+        assert($where instanceof Andx);
+
+        $parameters = $visitor->getParameters();
+
+        return [$where, $parameters];
     }
 
     /**