+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * refactor: simplify `KeysetExpressionCalculator`
 * chore: cleanup
 * feat: native query adapter now supports row values method
+* feat: query builder adapter now supports row values method
 
 # 0.15.1